FinanceCategoryAdd - Проведено кодревью - Добавлен цвет категориям

+++ app/src/main/java/com/example/appcash/data/vo/FinanceCategoryVO.kt
@@ -2,5 +2,5 @@ package com.example.appcash.data.vo
 
 data class FinanceCategoryVO(
     val name: String,
-    val color: Long
+    val color: Int
 )
